+++ Request/UploadFile.hpp
@@ -6,7 +6,7 @@
 /*   By: oait-laa <oait-laa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/03 10:58:22 by oait-laa          #+#    #+#             */
-/*   Updated: 2025/01/10 11:12:46 by oait-laa         ###   ########.fr       */
+/*   Updated: 2025/01/11 10:36:49 by oait-laa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ class UploadFile {
         std::string filename;
         std::string type;
         std::string tmpContent;
-        size_t expectedBytes;
+        long long expectedBytes;
         bool firstRead;
         bool state;
         std::string boundary;
@@ -38,7 +38,7 @@ class UploadFile {
         std::ofstream* getFd();
         std::string getFilename();
         std::string getTmpContent();
-        size_t getExpectedBytes();
+        long long getExpectedBytes();
         std::string getType();
         bool isFirstRead();
         bool getState();
@@ -47,7 +47,7 @@ class UploadFile {
         // Setters
         void setFilename(std::string name);
         void setTmpContent(std::string tmp);
-        void setExpectedBytes(size_t bytes);
+        void setExpectedBytes(long long bytes);
         void setType(std::string n_type);
         void setState(bool s);
         void setFirstRead(bool s);
